# Optimize writing optional columns (part 3)