--- a/level_amd64.go
+++ b/level_amd64.go
@@ -1,0 +1,6 @@
+//go:build !purego
+
+package parquet
+
+//go:noescape
+func memset(dst []byte, src byte)