--- a/level_amd64.s
+++ b/level_amd64.s
@@ -7,7 +7,6 @@ TEXT ·memset(SB), NOSPLIT, $0-32
     MOVQ dst+0(FP), AX
     MOVQ dst+8(FP), BX
     MOVBQZX src+24(FP), CX
-    XORQ SI, SI
 
     CMPQ BX, $8
     JBE test
@@ -18,6 +17,7 @@ TEXT ·memset(SB), NOSPLIT, $0-32
     CMPB ·hasAVX2(SB), $0
     JE init8
 
+    XORQ SI, SI
     MOVQ BX, DX
     SHRQ $6, DX
     SHLQ $6, DX
@@ -33,23 +33,22 @@ loop64:
     VMOVDQU Y0, -32(AX)(BX*1)
     VZEROUPPER
     RET
+
 init8:
     MOVQ $0x0101010101010101, R8
     IMULQ R8, CX
-
-    MOVQ BX, DX
-    SHRQ $3, DX
-    SHLQ $3, DX
 loop8:
-    MOVQ CX, (AX)(SI*1)
-    ADDQ $8, SI
-    CMPQ SI, DX
-    JNE loop8
-    JMP test
+    MOVQ CX, -8(AX)(BX*1)
+    SUBQ $8, BX
+    CMPQ BX, $8
+    JAE loop8
+    MOVQ CX, (AX)
+    RET
+
 loop:
-    MOVB CX, (AX)(SI*1)
-    INCQ SI
+    MOVB CX, -1(AX)(BX*1)
+    DECQ BX
 test:
-    CMPQ SI, BX
+    CMPQ BX, $0
     JNE loop
     RET