--- a/level_purego.go
+++ b/level_purego.go
@@ -1,0 +1,9 @@
+//go:build purego || !amd64
+
+package parquet
+
+func memset(dst []byte, src byte) {
+	for i := range dst {
+		dst[i] = src
+	}
+}