Based on https://github.com/segmentio/parquet-go/pull/206, this PR adds optimizations for the `appendLevel` function employed to write definition and repetition levels.

Most of the time of this function was spent on a loop assigning single-byte level values:
```go
fill := levels[i:]
for i := range fill {
    fill[i] = value
}
```
I replaced this statement with a `memset` function (similar to the C stdlib).

The following snapshot of benchmark shows the intended effect of the code change:
```
name                                      old time/op  new time/op  delta
GenericBuffer/optionalInt32Column/go1.18  1.70µs ± 0%  1.53µs ± 1%   -9.96%  (p=0.000 n=8+8)

name                                      old row/s    new row/s    delta
GenericBuffer/optionalInt32Column/go1.18    589M ± 0%    654M ± 1%  +11.06%  (p=0.000 n=8+8)
```